fix: lending staking issue
