+++ apps/admins/src/server/api/processDispense.ts
@@ -1,177 +1,158 @@
-// import { ethers, utils } from "ethers";
-// import { ObjectId } from "mongodb";
-// import Web3 from "web3";
+import { Contract, providers, utils, Wallet } from "ethers";
+import { ObjectId } from "mongodb";
 
-// import { env } from "@/env.mjs";
+import { env } from "@/env.mjs";
 
-// import { getDB } from "../../cron/db";
-// import { JurisEscrowAbi } from "../../cron/JurisEscrow.abi";
-// import { jurisfundAbi } from "../../cron/JurisFund.abi";
+import { getDB } from "../../cron/db";
+import { JurisEscrowAbi } from "../../cron/JurisEscrow.abi";
+import { jurisfundAbi } from "../../cron/JurisFund.abi";
 
-// // MongoDB and Ethereum configuration
-// const safePrivateKey = env.SAFE_PRIVATE_KEY; // Should be securely managed
-// const jusdc = "0xD857A3CD0AF9Ab5e1c3298A44A81A18754161DAc";
-// const safeAddress = "0x5Ddf646e7beC68243cFbB61bB6E90c826c6F5CAD";
-// const jurisfundContractAddress = "0x2FDbD499ff0ACE66a9884572c88d7bb899118336";
+const SAFE_PRIVATE_KEY: string = env.SAFE_PRIVATE_KEY;
+const RPC_URL_FUJI: string = env.RPC_URL_FUJI;
+const SAFE_ADDRESS = "0x5Ddf646e7beC68243cFbB61bB6E90c826c6F5CAD";
+const JUSDC_ADDRESS = "0xD857A3CD0AF9Ab5e1c3298A44A81A18754161DAc";
+const JURIS_FUND_ADDRESS = "0x2FDbD499ff0ACE66a9884572c88d7bb899118336";
 
-// // Set up Web3
-// const web3 = new Web3(new Web3.providers.HttpProvider(process.env["RPC_PROVIDER_URL"] ?? ""));
-// const db = await getDB();
-// const collection = db.collection("BorrowersTable");
+interface Borrower {
+  id: string;
+  userId: string;
+  firstName: string;
+  lastName: string;
+  email: string;
+  phoneNumber: string;
+  stateLiveIn: string;
+  typeOfCase: string;
+  alreadyWorkingWithAttorney: string;
+  lawFirmName: string;
+  lawyerName: string;
+  lawyerEmail?: string;
+  lawyerWalletAddress?: string;
+  expectedSettlementAmount: string;
+  applicationFillingDate: Date;
+  caseNumber: string;
+  applicationStatus: string;
+  loanAmount?: string;
+  walletAddress: string;
+  fixedAPR?: string;
+  loanIssuanceDate?: Date;
+  accruedInterest?: string;
+  payoffDate?: Date;
+  transactionHash?: string;
+  escrowAddress?: string;
+}
 
-// async function fetchBorrowers(): Promise<Borrower[]> {
-//   try {
-//     const borrowers = await collection
-//       .find({ applicationStatus: "Approved", loanInssuanseDate: null })
-//       .toArray();
-//     return borrowers as unknown as Borrower[];
-//   } catch (error) {
-//     console.error(error);
-//     throw error;
-//   }
-// }
+const provider = new providers.JsonRpcProvider(RPC_URL_FUJI);
+const signer = new Wallet(SAFE_PRIVATE_KEY, provider);
+const jurisFundContract = new Contract(JURIS_FUND_ADDRESS, jurisfundAbi, signer);
 
-// // Function to sign and send transaction
-// async function signAndSendTransaction(plaintiff: string, amount: string): Promise<string> {
-//   const contract = new web3.eth.Contract(contractABI, contractAddress);
-//   const account = web3.eth.accounts.privateKeyToAccount(privateKey);
-//   web3.eth.accounts.wallet.add(account);
+const db = await getDB();
 
-//   // Check if contract.methods.dispense exists before invoking encodeABI
-//   // eslint-disable-next-line @typescript-eslint/strict-boolean-expressions
-//   if (contract.methods.dispense) {
-//     const data = contract.methods.dispense(tokenAddress, plaintiff, BigInt(amount)).encodeABI();
+const fetchBorrowers = async () => {
+  const collection = db.collection("BorrowersTable");
 
-//     const transaction = {
-//       to: contractAddress,
-//       data: data,
-//       gas: await web3.eth.estimateGas({ to: contractAddress, data: data }),
-//       gasPrice: await web3.eth.getGasPrice(),
-//     };
+  try {
+    const borrowers = await collection
+      .find({ applicationStatus: "Approved", loanIssuanceDate: null })
+      .toArray();
+    return borrowers as unknown as Borrower[];
+  } catch (error) {
+    console.error(error);
+    throw error;
+  }
+};
 
-//     const signedTransaction = await web3.eth.accounts.signTransaction(transaction, privateKey);
-//     const receipt = await web3.eth.sendSignedTransaction(signedTransaction.rawTransaction);
-//     return receipt.transactionHash as unknown as string;
-//   } else {
-//     throw new Error("dispense method is not defined on the contract");
-//   }
-// }
+const updateBorrowerData = async (
+  id: string,
+  transactionHash: string,
+  newEscrowAddress: string,
+): Promise<void> => {
+  const collection = db.collection("BorrowersTable");
 
-// // Function to update the borrower in MongoDB
-// async function updateBorrower(
-//   id: string,
-//   transactionHash: string,
-//   newEscrowAddress: string,
-// ): Promise<void> {
-//   try {
-//     const now = new Date();
-//     const objectId = new ObjectId(id);
-//     await collection.updateOne(
-//       { _id: objectId },
-//       {
-//         $set: {
-//           loanInssuanseDate: now,
-//           transactionHash: transactionHash,
-//           escrowAddress: newEscrowAddress,
-//         },
-//       },
-//     );
-//   } catch (error) {
-//     console.error("Error updating borrower:", error);
-//     throw error;
-//   }
-// }
+  try {
+    const now = new Date();
+    const objectId = new ObjectId(id);
+    await collection.updateOne(
+      { _id: objectId },
+      {
+        $set: {
+          loanIssuanceDate: now,
+          transactionHash: transactionHash,
+          escrowAddress: newEscrowAddress,
+        },
+      },
+    );
+  } catch (error) {
+    console.error("Error updating borrower:", error);
+    throw error;
+  }
+};
 
-// // Main processBorrowers function
-// export async function processBorrowers(): Promise<void> {
-//   try {
-//     const borrowers = await fetchBorrowers();
+const handle = async () => {
+  try {
+    const borrowers = await fetchBorrowers();
 
-//     for (const borrower of borrowers) {
-//       try {
-//         const transactionHash = await signAndSendTransaction(
-//           borrower.walletAddress,
-//           borrower.loanAmount ?? "",
-//         );
+    for (const borrower of borrowers) {
+      if (borrower.loanAmount == null) return;
+      borrower.loanAmount = utils.parseUnits(borrower.loanAmount, 6).toString();
 
-//         const newEscrowAddress = await getInitializationDataAndDeployEscrow(
-//           borrower.loanAmount ?? "",
-//           borrower.fixedAPY ?? "",
-//           borrower.walletAddress,
-//           borrower.lawyerWalletAddress ?? "",
-//           "0x5Ddf646e7beC68243cFbB61bB6E90c826c6F5CAD", // Privileged address - change it to env vars
-//           "0xD857A3CD0AF9Ab5e1c3298A44A81A18754161DAc", // JUSDC address - change it to env vars
-//         );
+      try {
+        const transactionHash = await dispenseLoan(borrower.walletAddress, borrower.loanAmount);
 
-//         console.log({ newEscrowAddress });
+        const { calldata, salt } = getEscrowInitializerAndSalt(borrower);
 
-//         await updateBorrower(borrower.id, transactionHash, newEscrowAddress);
-//       } catch (error) {
-//         console.error(`Error processing borrower ${borrower.id}:`, error);
-//         // Decide whether to continue or stop
-//       }
-//     }
-//   } catch (error) {
-//     console.error("Error fetching borrowers:", error);
-//     throw error;
-//   }
-// }
+        const escrowAddress = await deployEscrow(calldata, salt);
 
-// interface Borrower {
-//   id: string;
-//   userId: string;
-//   firstName: string;
-//   lastName: string;
-//   email: string;
-//   phoneNumber: string;
-//   stateLiveIn: string;
-//   typeOfCase: string;
-//   alreadyWorkingWithAttorney: string;
-//   lawFirmName: string;
-//   lawyerName: string;
-//   lawyerEmail?: string;
-//   lawyerWalletAddress?: string;
-//   expectedSettlementAmount: string;
-//   applicationFillingDate: Date;
-//   caseNumber: string;
-//   applicationStatus: string;
-//   loanAmount?: string;
-//   walletAddress: string;
-//   fixedAPY?: string;
-//   loanInssuanseDate?: Date;
-//   accruedInterest?: string;
-//   payoffDate?: Date;
-// }
+        await updateBorrowerData(borrower.id, transactionHash, escrowAddress);
+      } catch (error) {
+        console.error(`Error processing borrower ${borrower.id}:`, error);
+        // Decide whether to continue or stop
+      }
+    }
+  } catch (error) {
+    console.error("Error fetching borrowers:", error);
+    throw error;
+  }
+};
 
-// const getInitializationDataAndDeployEscrow = async (
-//   borrowerLoanAmount: string,
-//   apr: string,
-//   borrowerAddress: string,
-//   lawyerWalletAddress: string,
-//   safeAddress: string,
-//   usdcAddress: string,
-// ) => {
-//   const iface = new utils.Interface(JurisEscrowAbi);
+const getEscrowInitializerAndSalt = (data: Borrower) => {
+  const iface = new utils.Interface(JurisEscrowAbi);
 
-//   const callData = iface.encodeFunctionData("initialize", [
-//     borrowerLoanAmount, // loan amount
-//     apr, // the loan
-//     borrowerAddress, // borrowers address
-//     lawyerWalletAddress,
-//     safeAddress, // the privileged address
-//     usdcAddress, // JUSDC address
-//   ]);
+  const calldata = iface.encodeFunctionData("initialize", [
+    data.loanAmount,
+    data.fixedAPR,
+    data.walletAddress,
+    data.lawyerWalletAddress,
+    SAFE_ADDRESS,
+    JUSDC_ADDRESS,
+  ]);
 
-//   const provider = new ethers.providers.JsonRpcProvider(env.RPC_URL_FUJI, 1);
-//   const signer = new ethers.Wallet(env.SAFE_PRIVATE_KEY, provider);
-//   const jurisFundContract = new ethers.Contract(jurisfundContractAddress, JurisEscrowAbi, signer);
+  const salt = utils.keccak256(calldata);
 
-//   const escrowAddress = await jurisFundContract.deployEscrow(
-//     callData,
-//     "0xef50095700000000000000000000000000000000000000000000000000000000",
-//   );
+  return { calldata, salt };
+};
 
-//   // const data = jurisFundContract.methods.(tokenAddress, plaintiff, BigInt(amount)).encodeABI();
+const deployEscrow = async (calldata: string, salt: string): Promise<string> => {
+  const tx: providers.TransactionResponse = await jurisFundContract.deployEscrow(calldata, salt);
+  const receipt: providers.TransactionReceipt = await tx.wait();
 
-//   return escrowAddress as string;
-// };
+  const event = receipt.logs
+    .filter((log) => log.logIndex === 1)
+    .map((log) => jurisFundContract.interface.parseLog(log))
+    .find((item) => item.name === "EscrowCreated");
+
+  const proxyAddress: string = event?.args.proxy as string;
+
+  return proxyAddress;
+};
+
+const dispenseLoan = async (plaintiffAddress: string, loanAmount: string): Promise<string> => {
+  const tx: providers.TransactionResponse = await jurisFundContract.dispense(
+    JUSDC_ADDRESS,
+    plaintiffAddress,
+    loanAmount,
+  );
+  const receipt: providers.TransactionReceipt = await tx.wait();
+  return receipt.transactionHash;
+};
+
+export { handle };
